requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
pydantic>=2.8,<3.0
python-multipart>=0.0.9,<1.0
pypdf>=5.1,<6.0
python-docx>=1.1,<2.0
